+++ release_notes.md
@@ -17,6 +17,7 @@ PREESM Changelog
 
 ### Bug fix
 * Fix workflow task "org.ietr.preesm.pimm.algorithm.checker.periods.PeriodsPreschedulingChecker";
+* Fix workflow task "pisdf-export" (adds graph copy to avoid side-effects);
 * Fix codegen: special actor ports order is now preserved during codegen;